chore: Move asset-like folders below assets/
