+++ .gitignore
@@ -16,6 +16,6 @@ personal-setup
 /resources
 acceptance/acceptance.test
 .vscode
-statik/
-statikWebAssets/
-statikWebViews/
+assets/statik/
+assets/statikWebAssets/
+assets/statikWebViews/
