+++ Makefile
@@ -73,21 +73,21 @@ getstatik:
 	( [ -x "$$(command -v statik)" ] || go get github.com/rakyll/statik@v0.1.7 )
 
 update_registry:
-	helm package ./assets/container-registry/chart/container-registry/ -d embedded-files
+	helm package ./assets/container-registry/chart/container-registry/ -d assets/embedded-files
 
 update_google_service_broker:
 	@./scripts/update-google-service-broker.sh
 
 update_tekton:
-	mkdir -p embedded-files/tekton
-	wget https://storage.googleapis.com/tekton-releases/pipeline/previous/v0.19.0/release.yaml -O embedded-files/tekton/pipeline-v0.19.0.yaml
-	wget https://storage.googleapis.com/tekton-releases/triggers/previous/v0.11.1/release.yaml -O embedded-files/tekton/triggers-v0.11.1.yaml
-	wget https://github.com/tektoncd/dashboard/releases/download/v0.11.1/tekton-dashboard-release.yaml -O embedded-files/tekton/dashboard-v0.11.1.yaml
+	mkdir -p assets/embedded-files/tekton
+	wget https://storage.googleapis.com/tekton-releases/pipeline/previous/v0.19.0/release.yaml -O assets/embedded-files/tekton/pipeline-v0.19.0.yaml
+	wget https://storage.googleapis.com/tekton-releases/triggers/previous/v0.11.1/release.yaml -O assets/embedded-files/tekton/triggers-v0.11.1.yaml
+	wget https://github.com/tektoncd/dashboard/releases/download/v0.11.1/tekton-dashboard-release.yaml -O assets/embedded-files/tekton/dashboard-v0.11.1.yaml
 
 embed_files: getstatik
-	statik -m -f -src=./embedded-files
-	statik -m -f -src=./embedded-web-files/views -ns webViews -p statikWebViews
-	statik -m -f -src=./embedded-web-files/assets -ns webAssets -p statikWebAssets
+	statik -m -f -src=./assets/embedded-files -dest assets
+	statik -m -f -src=./assets/embedded-web-files/views -ns webViews -p statikWebViews -dest assets
+	statik -m -f -src=./assets/embedded-web-files/assets -ns webAssets -p statikWebAssets -dest assets
 
 help:
 	( echo _ _ ___ _____ ________ Overview ; epinio help ; for cmd in apps completion create-org delete help info install orgs push target uninstall ; do echo ; echo _ _ ___ _____ ________ Command $$cmd ; epinio $$cmd --help ; done ; echo ) | tee HELP
