+++ assets/embedded-files/app-ingress.yaml
