+++ assets/embedded-files/container-registry-0.1.0.tgz
