+++ assets/embedded-files/tekton/pipeline-v0.19.0.yaml
