+++ assets/embedded-files/tekton/staging.yaml
