+++ assets/embedded-web-files/assets/js/applications.js
