+++ assets/embedded-web-files/views/applications_index.html
