+++ assets/sample-app/htdocs/index.php
